delete incoming file from schema inference
